+++ frontend/src/pages/dashboard/play-battleground/components/common/Header.tsx
