+++ frontend/src/pages/dashboard/play-battleground/data/questions.ts
@@ -1,19 +1,6 @@
-import { useState, useEffect } from "react"
-import { useLocation, useNavigate } from "react-router-dom"
-import Header from "./Header"
-import VisibilityIcon from '@mui/icons-material/Visibility'
-import clockTimePressured from "../../../../assets/clockTimePressured.png"
+import { Question } from '../types';
 
-interface Question {
-    question: string;
-    correctAnswer: string;
-    options?: string[];  // Optional since identification won't have options
-    mode: 'Peaceful' | 'Time Pressured' | 'pvp';
-    difficulty?: 'easy' | 'average' | 'hard';  // Only for pvp mode
-    questionType: 'multiple-choice' | 'identification' | 'true-false';
-}
-
-const questionsData: Question[] = [
+export const questionsData: Question[] = [
     // Multiple Choice Questions (10 per mode)
     {
         question: "What does HTML stand for?",
@@ -618,358 +605,4 @@ const questionsData: Question[] = [
         mode: "Peaceful",
         questionType: "true-false"
     }
-]
-
-export default function MainFlashCardLayout() {
-    const [isFlipped, setIsFlipped] = useState(false)
-    const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0)
-    const [selectedAnswer, setSelectedAnswer] = useState<string | null>(null)
-    const [inputAnswer, setInputAnswer] = useState("")
-    const [showResult, setShowResult] = useState(false)
-    const [showNextButton, setShowNextButton] = useState(false)
-    const [score, setScore] = useState({ correct: 0, incorrect: 0 })
-    const location = useLocation()
-    const {
-        mode = 'Peaceful',
-        material = 'default',
-        selectedTypes = ['multiple-choice'],  // Updated from selectedQuestionTypes
-        timeLimit = null, // Add timeLimit with null default
-        difficulty = null, // Add difficulty with null default
-    } = location.state || {}
-    const navigate = useNavigate()
-    const [startTime] = useState(new Date())
-    const [correctCount, setCorrectCount] = useState(0)
-    const [incorrectCount, setIncorrectCount] = useState(0)
-    const [timeRemaining, setTimeRemaining] = useState<number | null>(null)
-    const [questionTimer, setQuestionTimer] = useState<number | null>(null);
-    const [timerProgress, setTimerProgress] = useState(100); // Progress percentage
-
-    // Add safety check for required props
-    if (!location.state) {
-        console.error('Missing required state properties')
-        navigate('/dashboard/home')
-        return null
-    }
-
-    // Filter questions based on game mode and question types
-    const filteredQuestions = questionsData.filter(question => {
-        if (mode === 'pvp') {
-            return question.mode === mode &&
-                question.difficulty === difficulty &&
-                selectedTypes.includes(question.questionType)
-        }
-        return question.mode === mode &&
-            selectedTypes.includes(question.questionType)
-    })
-
-    // Add safety check for empty filtered questions
-    if (filteredQuestions.length === 0) {
-        return <div>No questions available for this mode.</div>
-    }
-
-    const currentQuestion = filteredQuestions[currentQuestionIndex]
-
-    // Add safety check for undefined current question
-    if (!currentQuestion) {
-        return <div>Question not found.</div>
-    }
-
-    // Add this helper function to check if it's the last question
-    const isLastQuestion = currentQuestionIndex === filteredQuestions.length - 1;
-
-    // Initialize timer when component mounts or when moving to next question
-    useEffect(() => {
-        let timer: NodeJS.Timeout;
-
-        // Only start timer if in Time Pressured mode and not showing result
-        if (timeLimit && mode === "Time Pressured" && !showResult) {
-            setQuestionTimer(timeLimit);
-            setTimerProgress(100);
-
-            timer = setInterval(() => {
-                setQuestionTimer(prev => {
-                    if (prev === null || prev <= 0) {
-                        clearInterval(timer);
-                        // Auto submit wrong answer when time is up
-                        if (!showResult) {
-                            handleAnswerSubmit('');
-                            // Don't auto-advance, let user see the correct answer first
-                        }
-                        return 0;
-                    }
-                    // Calculate progress percentage
-                    const progress = (prev / timeLimit) * 100;
-                    setTimerProgress(progress);
-                    return prev - 1;
-                });
-            }, 1000);
-        }
-
-        return () => clearInterval(timer);
-    }, [currentQuestionIndex, timeLimit, showResult, mode]);
-
-    const handleAnswerSubmit = (answer: string) => {
-        if (showResult) return;
-
-        const isCorrect = answer.toLowerCase() === currentQuestion.correctAnswer.toLowerCase();
-
-        if (isCorrect) {
-            setCorrectCount(prev => prev + 1);
-        } else {
-            setIncorrectCount(prev => prev + 1);
-        }
-
-        setSelectedAnswer(answer);
-        setIsFlipped(true);
-        setShowResult(true);
-        setShowNextButton(true);
-    };
-
-    const handleNextQuestion = () => {
-        if (!isLastQuestion) {
-            setIsFlipped(false);
-
-            setTimeout(() => {
-                setCurrentQuestionIndex(prev => prev + 1);
-                setSelectedAnswer(null);
-                setInputAnswer("");
-                setShowResult(false);
-                setShowNextButton(false);
-                // Timer will automatically start for next question due to useEffect dependencies
-            }, 300);
-        } else {
-            handleGameComplete();
-        }
-    };
-
-    const handleGameComplete = () => {
-        const endTime = new Date();
-        const timeDiff = endTime.getTime() - startTime.getTime();
-
-        // Format time as mm:ss
-        const minutes = Math.floor(timeDiff / 60000);
-        const seconds = Math.floor((timeDiff % 60000) / 1000);
-        const timeSpent = `${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
-
-        navigate('/dashboard/study/session-summary', {
-            state: {
-                timeSpent,
-                correctCount,
-                incorrectCount,
-                mode,
-                material
-            }
-        });
-    }
-
-    const getButtonStyle = (option: string) => {
-        if (!showResult) {
-            return `border ${selectedAnswer === option ? 'border-[#4D18E8] border-2' : 'border-gray-800'}`
-        }
-
-        if (option === currentQuestion.correctAnswer) {
-            return 'border-[#52A647] border-2' // Correct answer
-        }
-
-        if (option === selectedAnswer && option !== currentQuestion.correctAnswer) {
-            return 'border-[#FF3B3F] border-2' // Wrong selected answer
-        }
-
-        return 'border border-gray-800' // Keep default border for unselected options
-    }
-
-    const renderQuestionContent = () => {
-        switch (currentQuestion.questionType) {
-            case 'multiple-choice':
-                return (
-                    <div className="grid grid-cols-2 lg:grid-cols-4 gap-4 w-full max-w-[1000px] mx-auto">
-                        {currentQuestion.options?.map((option, index) => (
-                            <button
-                                key={index}
-                                onClick={() => handleAnswerSubmit(option)}
-                                disabled={showResult}
-                                className={`h-[100px] w-full bg-transparent 
-                                    ${getButtonStyle(option)}
-                                    rounded-lg text-white hover:bg-gray-800/20 transition-colors
-                                    disabled:cursor-not-allowed px-4 text-center`}
-                            >
-                                {option}
-                            </button>
-                        ))}
-                    </div>
-                );
-
-            case 'identification':
-                return (
-                    <div className="w-full max-w-[500px] mt-3 mx-auto">
-                        <input
-                            type="text"
-                            value={inputAnswer}
-                            onChange={(e) => setInputAnswer(e.target.value)}
-                            onKeyPress={(e) => {
-                                if (e.key === 'Enter' && !showResult) {
-                                    handleAnswerSubmit(inputAnswer);
-                                }
-                            }}
-                            disabled={showResult}
-                            className={`w-full p-4 rounded-lg bg-transparent py-10 px-10 border-2 text-center
-                                ${showResult
-                                    ? inputAnswer.toLowerCase() === currentQuestion.correctAnswer.toLowerCase()
-                                        ? 'border-[#52A647]'
-                                        : 'border-[#FF3B3F]'
-                                    : 'border-gray-600'
-                                }
-                                text-white focus:outline-none placeholder:text-[#6F658D] placeholder:opacity-100`}
-                            placeholder="Type your answer here..."
-                        />
-                    </div>
-                );
-
-            case 'true-false':
-                return (
-                    <div className="flex gap-4 justify-center">
-                        {['true', 'false'].map((option) => (
-                            <button
-                                key={option}
-                                onClick={() => handleAnswerSubmit(option)}
-                                disabled={showResult}
-                                className={`h-[100px] w-[200px] bg-transparent 
-                                    ${getButtonStyle(option)}
-                                    rounded-lg text-white hover:bg-gray-800/20 transition-colors
-                                    disabled:cursor-not-allowed`}
-                            >
-                                {option.toUpperCase()}
-                            </button>
-                        ))}
-                    </div>
-                );
-        }
-    };
-
-    // Format time as mm:ss
-    const formatTime = (seconds: number | null) => {
-        if (seconds === null) return ''
-        const minutes = Math.floor(seconds / 60)
-        const remainingSeconds = seconds % 60
-        return `${minutes.toString().padStart(2, '0')}:${remainingSeconds.toString().padStart(2, '0')}`
-    }
-
-    // Modify the TimerDisplay component
-    const TimerDisplay = () => {
-        if (!questionTimer && questionTimer !== 0) return null;
-
-        return (
-            <div className="fixed bottom-0 left-0 w-full">
-                {/* Timer text with clock icon - Positioned at right side */}
-                <div className="absolute left-12 bottom-8 flex items-center gap-2">
-                    <img
-                        src={clockTimePressured}
-                        alt="Timer"
-                        className="w-6 h-6"  // Adjust size as needed
-                    />
-                    <div
-                        className={`text-2xl font-bold ${questionTimer <= 7
-                            ? 'text-red-500 animate-pulse'
-                            : 'text-white'
-                            }`}
-                    >
-                        {questionTimer}s
-                    </div>
-                </div>
-
-                {/* Progress bar - At the bottom */}
-                <div className="w-full h-2 bg-gray-700">
-                    <div
-                        className={`h-full transition-all duration-1000 ${questionTimer <= 7 ? 'bg-red-500' : 'bg-[#fff]'}`}
-                        style={{ width: `${timerProgress}%` }}
-                    />
-                </div>
-            </div>
-        );
-    };
-
-    return (
-        <div className="min-h-screen relative">
-            <Header
-                material={material}
-                mode={mode}
-                correct={correctCount}
-                incorrect={incorrectCount}
-            />
-            <main className="pt-24 px-4">
-                <div className="mx-auto max-w-[1200px] flex flex-col items-center gap-8 h-[calc(100vh-96px)] justify-center">
-                    {/* Question Card */}
-                    <div
-                        className="w-full max-w-[900px] h-[380px] mt-[-60px] bg-white rounded-lg p-8 cursor-pointer relative"
-                        onClick={() => !showResult && setIsFlipped(!isFlipped)}
-
-                        style={{
-                            perspective: "1000px",
-                            transformStyle: "preserve-3d",
-                            transition: "transform 0.6s",
-                            transform: isFlipped ? "rotateY(180deg)" : "rotateY(0deg)",
-                        }}
-                    >
-                        {/* Front of card (Question) */}
-                        <div
-                            className="absolute inset-0 backface-hidden flex flex-col items-center justify-center p-8 bg-white rounded-lg"
-                            style={{
-                                backfaceVisibility: "hidden",
-                                transition: "opacity 0.3s",
-                                opacity: isFlipped ? 0 : 1
-                            }}
-                        >
-                            <div className="flex-1 flex items-center justify-center">
-                                <p className="text-center text-black text-2xl max-w-[600px]">
-                                    {currentQuestion.question}
-                                </p>
-                            </div>
-
-                            {/* Reveal Answer Button */}
-                            <button
-                                className="absolute bottom-6 right-9 flex items-center space-x-2 text-[#4D18E8] 
-                                hover:text-[#4D18E8] transition-all duration-200 ease-in-out
-                                hover:scale-105 hover:font-bold transform "
-                            >
-                                <VisibilityIcon className="w-5 h-5 text-[#4D18E8]" />
-                                <span className="text-sm font-bold text-[#4D18E8]">REVEAL ANSWER</span>
-                            </button>
-                        </div>
-
-                        {/* Back of card (Answer) */}
-                        <div
-                            className="absolute inset-0 backface-hidden flex items-center justify-center bg-white rounded-lg"
-                            style={{
-                                backfaceVisibility: "hidden",
-                                transform: "rotateY(180deg)",
-                                transition: "opacity 0.3s",
-                                opacity: isFlipped ? 1 : 0
-                            }}
-                        >
-                            <p className="text-center text-black text-3xl font-bold">
-                                {currentQuestion.correctAnswer}
-                            </p>
-                        </div>
-                    </div>
-
-                    {/* Answer Section */}
-                    {renderQuestionContent()}
-
-                    {/* Next Button */}
-                    {showNextButton && (
-                        <button
-                            onClick={handleNextQuestion}
-                            className="mt-6 px-8 py-3 bg-[#4D18E8] text-white rounded-lg hover:bg-[#3A12B0] transition-colors"
-                        >
-                            {isLastQuestion ? 'Done' : 'Next Question'}
-                        </button>
-                    )}
-                </div>
-            </main>
-
-            {/* Add TimerDisplay component */}
-            {mode === "Time Pressured" && <TimerDisplay />}
-        </div>
-    )
-}
-
+];
